+++ src/main/java/com/github/mwttg/sjge/configuration/Configuration.java
@@ -1,0 +1,16 @@
+package com.github.mwttg.sjge.configuration;
+
+public record Configuration(GameWindowConfiguration gameWindowConfiguration,
+                            OpenGlConfiguration openGlConfiguration,
+                            ViewPortConfiguration viewPortConfiguration) {
+
+    public String prettyPrint() {
+        return """
+                Configuration
+                %s%s%s
+                """.formatted(
+                gameWindowConfiguration.prettyPrint(),
+                openGlConfiguration.prettyPrint(),
+                viewPortConfiguration.prettyPrint());
+    }
+}