+++ src/main/java/com/github/mwttg/sjge/graphics/camera/ProjectionMatrix.java
@@ -1,0 +1,4 @@
+package com.github.mwttg.sjge.graphics.camera;
+
+public class ProjectionMatrix {
+}