+++ src/main/java/com/github/mwttg/sjge/graphics/window/Configuration.java
@@ -1,7 +1,0 @@
-package com.github.mwttg.sjge.graphics.window;
-
-public record Configuration(String title,
-                            int width,
-                            int height,
-                            OpenGlConfiguration openGlConfiguration) {
-}