+++ src/main/java/com/github/mwttg/sjge/graphics/window/OpenGlConfiguration.java
@@ -1,8 +1,0 @@
-package com.github.mwttg.sjge.graphics.window;
-
-public record OpenGlConfiguration(int openGlMajorVersion,
-                                  int openGlMinorVersion,
-                                  boolean vsync,
-                                  boolean wireframe,
-                                  boolean backfaceCulling) {
-}